requirements.txt
Flask==2.3.3
qrcode==7.4
Pillow==10.0.1
reportlab==4.0.8
gunicorn==22.1.0
